add dry_run for run_submission() (#310)
For quick test, it might be in need of dry-run instead of just returning
after submission really take effect. For easily debug, `dry_run` is add
as a choice to make it possible to only put the input files onto
work_path but not really submitted.

While the feature requested and reviewed, it might be a question whether
a so-called dry-run should contain the step of uploading instead of just
print the file on the screen. Maybe more discussion in need?